% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ne.R
\name{Ne}
\alias{Ne}
\title{Evaluate effective population size (Ne).}
\usage{
Ne(nM, nF, AvgProgM = (nM + nF)/nM, AvgProgF = (nM + nF)/nF,
  VarProgM = AvgProgM, VarProgF = AvgProgF, c = nM/(nM + nF), Fis = 0)
}
\arguments{
\item{nM}{numeric, number of breeding males.}

\item{nF}{numeric, number of breeding females.}

\item{AvgProgM}{numeric, average number of progeny (family size) per male
(number of progeny that will contribute to the next generation!!!).}

\item{AvgProgF}{numeric, average number of progeny (family size) per female
(number of progeny that will contribute to the next generation!!!).}

\item{VarProgM}{numeric, average number of progeny (family size) per male
(number of progeny that will contribute to the next generation!!!).}

\item{VarProgF}{numeric, average number of progeny (family size) per female
(number of progeny that will contribute to the next generation!!!).}

\item{c}{numeric, fraction of males among breeding individuals.}

\item{Fis}{numeric, inbreeding coefficient (within (sub)-population) due to
the excess of matings between relatives over random mating expectation.}
}
\value{
Evaluated effective population size.
}
\description{
\code{Ne} evaluates effective population size given some demographic parameters.
}
\examples{
Ne(nM=100, nF=100)
Ne(nM=10, nF=100)
Ne(nM=10, nF=100, VarProgM=0)
}

