% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/System.R
\name{System}
\alias{System}
\title{Run system command with more control/reporting}
\usage{
System(Command, Wait = TRUE, CheckValue = 0, NoteStart = NULL,
  NoteEnd = NULL, NoteError = NULL, NoteDetail = NULL, ...)
}
\arguments{
\item{Command}{\code{command} in \code{\link{system}}.}

\item{Wait}{\code{wait} in \code{\link{system}}.}

\item{CheckValue}{numeric, check the command return value against this value.}

\item{NoteStart}{character, add some notes prior to starting the command.}

\item{NoteEnd}{character, add some notes after the command has been run.}

\item{NoteError}{character, add some notes only when the error happens.}

\item{NoteDetail}{character, add some detailed notes to all the notes.}

\item{...}{arguments passed to \code{\link{system}}}
}
\value{
As in \code{\link{system}}.
}
\description{
\code{System} runs system command via \code{system} and gives you more reporting.
}
\examples{
System(Command="ls -l")

}
\seealso{
\code{\link{system}}
}

